requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
transformers
Pillow